requirements.txt
certifi==2021.5.30
cftime==1.5.0
cycler==0.10.0
Cython==0.29.24
joblib==1.0.1
kiwisolver==1.3.2
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pyerfa==2.0.0
pyModeS==2.9
pymongo==3.12.0
pyparsing==2.4.7
pyrtlsdr==0.2.92
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pyzmq==22.3.0
Rx==3.2.0
scipy==1.7.1
six==1.16.0
tqdm==4.62.3
urllib3==1.26.7
xarray==0.15.1
